requirements.txt
Django
djangorestframework
drf-spectacular
drf-spectacular-sidecar
django-taggit
django-extensions
django-jazzmin
icecream